manager/requirements.txt
redis==2.10.6
pyzmq==17.0.0b3
